refactor: move Block and Model classes up into Milestones namespace
